Create a separate ScanComplete component
