+++ app/views/QRScan.js
@@ -1,54 +1,16 @@
 import React, { useEffect, useState } from 'react';
-import { useTranslation } from 'react-i18next';
-import {
-  BackHandler,
-  Dimensions,
-  ImageBackground,
-  StatusBar,
-  View,
-} from 'react-native';
-import { SvgXml } from 'react-native-svg';
+import { BackHandler, ImageBackground, StatusBar } from 'react-native';
 
 import BackgroundImage from './../assets/images/launchScreenBackground.png';
-import StateAtRisk from './../assets/svgs/stateAtRisk';
-import StateNoContact from './../assets/svgs/stateNoContact';
-import { Button } from '../components/Button';
-import { Typography } from '../components/Typography';
 import { Theme } from '../constants/themes';
 import { isValidCoordinates } from '../helpers/Location';
 import LocationServices from '../services/LocationService';
+import { ScanComplete } from './QRScan/ScanComplete';
 import { ScanInProgress } from './QRScan/ScanInProgress';
+import { StateEnum } from './QRScan/StateEnum';
 import { styles } from './QRScan/style';
 
-const StateEnum = {
-  DEFAULT: 0,
-  SCAN_SUCCESS: 1,
-  SCAN_FAIL: 2,
-  SCAN_IN_PROGRESS: 3,
-};
-
-const StateIcon = ({ status, size }) => {
-  let icon;
-  switch (status) {
-    case StateEnum.DEFAULT:
-      icon = null;
-      break;
-    case StateEnum.SCAN_SUCCESS:
-      icon = StateNoContact;
-      break;
-    case StateEnum.SCAN_FAIL:
-      icon = StateAtRisk;
-      break;
-  }
-  return (
-    <SvgXml xml={icon} width={size ? size : 80} height={size ? size : 80} />
-  );
-};
-
-const height = Dimensions.get('window').height;
-
 export const QRScanScreen = ({ navigation, route }) => {
-  const { t } = useTranslation();
   const [currentState, setcurrentState] = useState(StateEnum.DEFAULT);
 
   const onNavigate = () => {
@@ -112,30 +74,12 @@ export const QRScanScreen = ({ navigation, route }) => {
     }
   };
 
-  const getMainText = () => {
-    switch (currentState) {
-      case StateEnum.SCAN_SUCCESS:
-        return (
-          <Typography style={styles.mainTextBelow} use='headline2'>
-            {t('qr.successful_title')}
-          </Typography>
-        );
-      case StateEnum.SCAN_FAIL:
-        return (
-          <Typography style={styles.mainTextBelow} use='headline2'>
-            {t('qr.fail_title')}
-          </Typography>
-        );
-    }
-  };
+  const isScanComplete =
+    currentState === StateEnum.SCAN_SUCCESS ||
+    currentState === StateEnum.SCAN_FAIL;
 
-  const getSubText = () => {
-    switch (currentState) {
-      case StateEnum.SCAN_SUCCESS:
-        return t('qr.successful_subtitle');
-      case StateEnum.SCAN_FAIL:
-        return t('qr.fail_subtitle');
-    }
+  const exitQRScan = () => {
+    navigation.navigate('Main', {});
   };
 
   return (
@@ -146,37 +90,13 @@ export const QRScanScreen = ({ navigation, route }) => {
           backgroundColor='transparent'
           translucent
         />
-        {(currentState === StateEnum.SCAN_SUCCESS ||
-          currentState === StateEnum.SCAN_FAIL) && (
-          <View style={styles.iconContainer}>
-            <StateIcon size={height} status={currentState} />
-          </View>
-        )}
-        {(currentState === StateEnum.SCAN_SUCCESS ||
-          currentState === StateEnum.SCAN_FAIL) && (
-          <View style={styles.mainContainer}>
-            <View style={styles.content}>
-              {getMainText()}
-              <Typography style={styles.subheaderText} use='body1'>
-                {getSubText()}
-              </Typography>
-              <View style={styles.buttonContainer}>
-                <Button
-                  label={t('qr.exit')}
-                  onPress={() => {
-                    navigation.navigate('Main', {});
-                  }}
-                />
-              </View>
-            </View>
-          </View>
+        {isScanComplete && (
+          <ScanComplete currentState={currentState} onExit={exitQRScan} />
         )}
         {currentState === StateEnum.SCAN_IN_PROGRESS && (
           <ScanInProgress
             onScanSuccess={onScanSuccess}
-            onScanCancel={() => {
-              navigation.navigate('Main', {});
-            }}
+            onScanCancel={exitQRScan}
           />
         )}
       </ImageBackground>
