+++ app/views/QRScan/StateEnum.js
@@ -1,0 +1,6 @@
+export const StateEnum = {
+  DEFAULT: 0,
+  SCAN_SUCCESS: 1,
+  SCAN_FAIL: 2,
+  SCAN_IN_PROGRESS: 3,
+};
